+++ src/napari_blob_detection/__init__.py
@@ -1,4 +1,5 @@
 __version__ = "0.0.1"
 
+from ._detect import difference_of_gaussian, laplacian_of_gaussian
 from ._points_to_labels import points_to_labels
 from ._widget import difference_of_gaussian, laplacian_of_gaussian, detect_blobs_widget
